pyproject.toml
[project]
name = "wavjepa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.2.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "braceexpand==0.1.7",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "docker-pycreds==0.4.0",
    "einops==0.8.1",
    "filelock==3.18.0",
    "fonttools==4.58.0",
    "frozenlist==1.6.0",
    "fsspec==2025.5.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "grpcio==1.71.0",
    "hydra-core==1.3.2",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "kiwisolver==1.4.8",
    "lightning==2.5.1.post0",
    "lightning-utilities==0.14.3",
    "markdown==3.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mpmath==1.3.0",
    "multidict==6.4.4",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "omegaconf==2.3.0",
    "packaging==24.2",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "propcache==0.3.1",
    "protobuf==6.31.0",
    "psutil==7.0.0",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.4.0",
    "pytorch-metric-learning==2.8.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "sentry-sdk==2.29.1",
    "setproctitle==1.3.6",
    "six==1.17.0",
    "smmap==5.0.2",
    "soundfile==0.13.1",
    "sympy==1.14.0",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "threadpoolctl==3.6.0",
    "torch==2.7.0",
    "torchaudio==2.7.0",
    "torchmetrics==1.7.1",
    "tqdm==4.67.1",
    "triton==3.3.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "wandb==0.19.11",
    "webdataset==0.2.111",
    "werkzeug==3.1.3",
    "yarl==1.20.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.ruff.lint]
extend-select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
